Jekyll had permalink, pelican doesn't.
Permalink in jekyll allowed to change the whole URL for each post,
pelican doesn't have this functionality. You can specify 'slug' for each
post, but with pelicanconf.py having ARTICLE_SAVE_AS = '{slug}'.html means that it will for sure have a suffix which we don't want.

Best solution seems to change the file name to have full slug as we like
it.